moved _pickle_args to Execution.py as cluster_runnable is the only function using it
